+++ tests/testthat/test-MISSSBM-fit-with-covariates.R
@@ -11,7 +11,7 @@ control <- list(threshold = 1e-2, maxIter = 50, fixPointIter = 5, trace = TRUE)
 test_that("missSBM with covariates and dyad sampling works", {
 
   ## Consistency
-  tol_truth <- .2
+  tol_truth <- .4
   tol_ARI   <- .5
 
   sampler_undirected_cov$rNetwork(store = TRUE)
@@ -87,7 +87,7 @@ test_that("missSBM with covariates and dyad sampling works", {
 test_that("miss SBM with covariates and node sampling works", {
 
   ## Consistency
-  tol_truth <- .2
+  tol_truth <- .3
   tol_ARI   <- .5
 
   sampler_undirected_cov_node$rNetwork(store = TRUE)