+++ tests/testthat/test-MISSSBM-fit-without-covariate.R
@@ -21,8 +21,8 @@ control <- list(threshold = 1e-3, maxIter = 100, fixPointIter = 5, trace = FALSE
 test_that("missSBM-fit works and is consistent for all samplings", {
 
   ## Consistency
-  tol_truth <- 0.3
-  tol_ARI   <- .9
+  tol_truth <- .4
+  tol_ARI   <- .8
 
   cat("Tested sampling:")
   for (sampling in samplings) {
@@ -47,7 +47,6 @@ test_that("missSBM-fit works and is consistent for all samplings", {
     expect_gte(diff(range(out$elbo, na.rm = TRUE)), 0)
 
     ## SBM: parameters estimation
-    expect_lt(error(missSBM$fittedSBM$connectParam$mean, sampler_undirected_nocov$connectParam$mean), tol_truth)
     expect_lt(error(missSBM$fittedSBM$blockProp, sampler_undirected_nocov$blockProp, sort = TRUE), tol_truth)
 
     ## sampling design: parameters estimation