+++ src/nlopt_wrapper.h
@@ -4,7 +4,6 @@
 #pragma once
 
 #include <Rcpp.h>
-// #include <nlopt.h>
 #include <nloptrAPI.h>
 
 #include <memory>      // unique_ptr