+++ NEWS.md
@@ -1,4 +1,8 @@
-# missSBM 1.0.0
+# missSBM dev version
+
+ongoing...
+
+# missSBM 1.0.0 [2021-05-25]
 
   - now rely on future_lapply  for parallel computing (plan to be set by the user)
   - faster model exploration (used to be called 'smoothing'), now integrated by default in estimateMissSBM